refactor(ui): TMS-2209 add index files
Added more index files for the barrel imports.
