+++ frontend/src/components/CreateTenantDialog.vue
@@ -4,7 +4,7 @@ import { VForm } from 'vuetify/components'
 
 import ButtonPrimary from '@/components/ui/ButtonPrimary.vue'
 import ButtonSecondary from '@/components/ui/ButtonSecondary.vue'
-import { useAuthStore } from '@/stores/useAuthStore'
+import { useAuthStore } from '@/stores'
 import { MINISTRIES } from '@/utils/constants'
 
 // Auto-bound v-model from parent
