+++ frontend/src/components/layout/AppNotifications.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useNotification } from '@/composables/useNotification'
+import { useNotification } from '@/composables'
 
 const { notifications, removeNotification } = useNotification()
 </script>
