+++ frontend/src/types/index.ts
@@ -1,0 +1,3 @@
+export type { IdirSearchParameters } from './IdirSearchParameters'
+export type { Notification } from './Notification'
+export type { NotificationType } from './NotificationType'
